[dhcpmon] Filter DHCP O/A Messages of Neighboring Vlans (#4469)
* [dhcpmon] Filter DHCP O/A Messages of Neighboring Vlans

This code fixes a bug where two or more vlans exist. Cross contamination
happens for DHCP packets Offer/Ack when received on shared northbound links.
The code filters out those packet based on dst IP equal Vlan loopback IP.

+++ src/dhcpmon/src/dhcp_mon.h
@@ -40,7 +40,7 @@ void dhcp_mon_shutdown();
  *
  * @return 0 upon success, otherwise upon failure
  */
-int dhcp_mon_start(int snaplen);
+int dhcp_mon_start(size_t snaplen);
 
 /**
  * @code dhcp_mon_stop();
